+++ CHANGES.rst
@@ -13,6 +13,7 @@ Changelog
   - every items, blank or not are created from an item template, this avoid use of `portal_factory`;
   - a special not removale `Default item template` is added in the `MeetingConfig` and is used as basis for creating a blank item;
   - parameter `MeetingConfig.itemCreatedOnlyUsingTemplate` is removed, deactivating the `Default item template` is the equivalent.
+- A MeetingConfig may be removed even if still containing items (recurring items, item templates), only real items are now considered
 
 4.1.24 (2020-05-08)
 -------------------